+++ packages/mint-components/stencil.config.ts
@@ -26,6 +26,9 @@ const useGrapesjs: OutputTarget = grapesJsOutput({
 
 export const config: Config = {
   namespace: "mint-components",
+  devServer: {
+    startupTimeout: 0,
+  },
   globalScript: "src/global/global.ts",
   globalStyle: "src/global/global.css",
   buildEs5: process.env.NODE_ENV !== "dev",