+++ packages/mint-components/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.2] - 2022-03-08
+
+### Changed
+
+- Changed Components:
+  - \<sqm-portal-login>
+    - Enforce that nextPage redirects are relative
+  - \<sqm-portal-reset-password>
+    - Enforce that nextPage redirects are relative
+  - \<sqm-portal-verify-email>
+    - Enforce that nextPage redirects are relative
+
 ## [1.5.1] - 2022-03-03
 
 ### Added
@@ -373,9 +385,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - \<sqm-popup-container>
   - \<sqm-stencilbook>
 
-[unreleased]: https://github.com/saasquatch/program-tools/compare/mint-components@1.5.1...HEAD
+[unreleased]: https://github.com/saasquatch/program-tools/compare/mint-components@1.5.2...HEAD
+[1.5.2]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.5.2
 [1.5.1]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.5.1
-[unreleased]: https://github.com/saasquatch/program-tools/compare/mint-components@1.5.0...HEAD
 [1.5.0]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.5.0
 [1.4.2]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.4.2
 [1.4.1]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.4.1